before api

+++ src/main/java/com/example/virtiverse/repository/OurUserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OurUserRepo extends JpaRepository<User, Integer> {
+public interface OurUserRepo extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
     boolean existsByEmail(String email);
